requirements.txt
pillow~=10.0.1
ffmpeg~=4.2.3
diffusers~=0.18.2
transformers~=4.32.1
torch
safetensors~=0.4.0